+++ docs/changes.rst
@@ -13,6 +13,10 @@ v0.5
 ----
 - Support more array-like data types for tabular data and list-like data types
   for text data. :pr:`179` by `Francesco Cariaggi`_.
+- Add an option `use_intelex` to :func:`skops.hub_utils.init` which, when
+  enabled, will result in the Hugging Face inference API running with Intel's
+  scikit-learn intelex library, which can accelerate inference times. :pr:`267`
+  by `Benjamin Bossan`_.
 
 v0.4
 ----